pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arango-entity-resolution"
dynamic = ["version"]
description = "A production-ready entity resolution system for ArangoDB using record blocking, graph algorithms, and AI."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Entity Resolution Team" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-arango>=8.0.0",
    "pandas>=2.1.4",
    "numpy>=1.24.4",
    "python-Levenshtein>=0.23.0",
    "jellyfish>=1.0.3",
    "fuzzywuzzy>=0.18.0",
    "scikit-learn>=1.3.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
onnx = [
    "onnxruntime>=1.16.0",
]
sparse = [
    "scipy>=1.11.0",
]
mcp = [
    "mcp>=1.0.0",
]
llm = [
    "litellm>=1.0.0",
]
test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.0.0",
]
dev = [
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[tool.hatch.version]
path = "src/entity_resolution/utils/constants.py"
pattern = "__version__ = \"(?P<version>[^\"]+)\""

[tool.hatch.build.targets.wheel]
packages = ["src/entity_resolution"]
# Ensure 'src' is treated as a source directory so the package is just 'entity_resolution'
[tool.hatch.build.targets.wheel.sources]
"src" = ""

[project.scripts]
arango-er = "entity_resolution.cli:main"
arango-er-demo = "entity_resolution.demo:run_presentation_demo"
arango-er-mcp = "entity_resolution.mcp.server:main"

[project.urls]
Homepage = "https://github.com/ArthurKeen/arango-entity-resolution"
Documentation = "https://github.com/ArthurKeen/arango-entity-resolution#readme"
Issues = "https://github.com/ArthurKeen/arango-entity-resolution/issues"
